refactor: film_recordとappreciationのimportを整理
