+++ api/base_model/film_record/create/film_record_create_model.py
@@ -1,7 +1,7 @@
 from base_model.film_record.create.appreciation.film_appreciation_create_model import FilmAppreciationCreateModel
 from base_model.film_record.create.film.film_create_model import FilmCreateModel
 from pydantic import BaseModel, Field, validator
-from src.domain.film_record.appreciation.appreciation_status_enum import AppreciationStatusEnum
+from src.domain.film_record.appreciation import AppreciationStatusEnum
 
 
 class FilmRecordCreateModel(BaseModel):
