+++ api/src/application/film_record/command/interface/film_record_command_application.py
@@ -1,9 +1,7 @@
 from abc import ABC, abstractmethod
 
 from base_model.film_record.create.film_record_create_model import FilmRecordCreateModel
-from src.domain.film_record.film_record_entity import FilmRecordEntity
-from src.domain.film_record.film_record_id_object import FilmRecordIdObject
-from src.domain.film_record.film_record_repository import IFilmRecordRepository
+from src.domain.film_record import FilmRecordEntity, FilmRecordIdObject, IFilmRecordRepository
 
 
 class IFilmRecordCommandApplication(ABC):
