+++ api/src/domain/film_record/film_record_repository.py
@@ -1,7 +1,6 @@
 from abc import ABC, abstractmethod
 
-from src.domain.film_record.film_record_entity import FilmRecordEntity
-from src.domain.film_record.film_record_id_object import FilmRecordIdObject
+from src.domain.film_record import FilmRecordEntity, FilmRecordIdObject
 
 
 class IFilmRecordRepository(ABC):
