+++ api/src/infrastructure/film_record/film_record_repository.py
@@ -1,9 +1,7 @@
 import os
 
 from notion_client import Client
-from src.domain.film_record.film_record_entity import FilmRecordEntity
-from src.domain.film_record.film_record_id_object import FilmRecordIdObject
-from src.domain.film_record.film_record_repository import IFilmRecordRepository
+from src.domain.film_record import FilmRecordEntity, FilmRecordIdObject, IFilmRecordRepository
 
 
 class ImplFilmRecordRepository(IFilmRecordRepository):
