+++ api/tests/domain/film_record/test_film_record_entity.py
@@ -1,16 +1,13 @@
 from datetime import date
 
 import pytest
-from src.domain.film_record.appreciation.appreciation_status_enum import AppreciationStatusEnum
-from src.domain.film_record.appreciation.film_appreciation_entity import FilmAppreciationEntity
-from src.domain.film_record.appreciation.film_appreciation_id_object import FilmAppreciationIdObject
+from src.domain.film_record import FilmRecordEntity, FilmRecordIdObject
+from src.domain.film_record.appreciation import AppreciationStatusEnum, FilmAppreciationEntity, FilmAppreciationIdObject
 from src.domain.film_record.film.film_entity import FilmEntity
 from src.domain.film_record.film.genre.film_genre_enum import FilmGenreEnum
 from src.domain.film_record.film.poster.film_poster_object import FilmPosterObject
 from src.domain.film_record.film.series.film_series_object import FilmSeriesObject
 from src.domain.film_record.film.tmdb_id_object import TmdbIdObject
-from src.domain.film_record.film_record_entity import FilmRecordEntity
-from src.domain.film_record.film_record_id_object import FilmRecordIdObject
 from src.domain.film_record.watch_medium.watch_medium_enum import WatchMediumEnum
